+++ src/SeekDeepPrograms/SeekDeepUtils/SeekDeepUtilsRunner.cpp
@@ -182,7 +182,7 @@ int SeekDeepUtilsRunner::dryRunQualityFiltering(
 	std::string qualWindow;
 	uint32_t qualityWindowLength;
 	uint32_t qualityWindowStep;
-	uint32_t qualityWindowThres;
+	uint8_t qualityWindowThres;
 	if (setUp.setOption(qualWindow, "--qualWindow", "SlidingQualityWindow")) {
 		seqUtil::processQualityWindowString(qualWindow, qualityWindowLength,
 				qualityWindowStep, qualityWindowThres);
